break out ProviderSizeSerializer
