+++ api/serializers/__init__.py
@@ -34,3 +34,4 @@ from .group_serializer import GroupSerializer
 from .volume_serializer import VolumeSerializer
 from .no_project_serializer import NoProjectSerializer
 from .project_serializer import ProjectSerializer
+from .provider_size_serializer import ProviderSizeSerializer
